Fix FileNotFoundError when library is loaded as ZIP archive
Removes an unnecessary *.Path.exists() check that broke compatibility with ZIP
archive libraries.

zipfile.Path.exists() returns False for the root of the archive, so the removed
check always returned False when the library was given as a .zip file.

This check turned out to be unnecessary in the first place, since the subsequent
call to .iterdir() will raise an appropriate error when the file does not exist.